refactor: rename auth package and remove unnecessary comments in AuthRepositoryFirestore
