+++ app/src/test/java/com/android/sample/model/authentication/AuthRepositoryFirestoreTest.kt
@@ -1,4 +1,4 @@
-package com.android.sample.model.auth
+package com.android.sample.model.authentication
 
 import android.os.Looper
 import androidx.test.core.app.ApplicationProvider
